+++ R/4.1-modSwap.R
@@ -28,11 +28,11 @@
 #' @param object An \code{ExprsBinary} object to mutate.
 #'
 #' @export
-setGeneric("swap",
-           function(object, ...) standardGeneric("swap")
+setGeneric("modSwap",
+           function(object, ...) standardGeneric("modSwap")
 )
 
-#' @describeIn swap A method to mutate \code{ExprsBinary} objects.
+#' @describeIn modSwap A method to mutate \code{ExprsBinary} objects.
 #'
 #' @param how A character string. The method used to mutate case subjects. Select from
 #'  "rp.1", "rp.2", "fp", "ng", or "tg".
@@ -44,7 +44,7 @@ setGeneric("swap",
 #'  appended to the \code{$mutated} column of the \code{@@annot} slot.
 #'
 #' @export
-setMethod("swap", "ExprsBinary",
+setMethod("modSwap", "ExprsBinary",
           function(object, how = "fp", percent = 10, theta = 1){
 
             if(percent < 1 | percent > 100){